extract ColorizationAdapter to its own file
